added a base case
we love recursion! But, actually, lexing and parsing should now work with a read in file for an extremely simple program.
The syntax for testing is
lex(filename, X),
program(X,Y).
which should return something such as
X = ['{','}', $]
Y = [].
X will be the returned list of words.
Y will be an empty list.
If false is returned, then either the lex or the parse failed. It's not very descriptive right now, but hey, it works.