Semantic Analysis fixes
All of the good code from the first project's test cases should now correctly pass/fail in semantic analysis. This is a big fix, so here are some of the changes:

child/2 can now be checked for scopes before they are closed (so we don't end up with ProLog going "oh shit, I can't find the parent of this scope, let's pretend we're using all of the other scopes we've made"). This was causing some false positives that didn't show up until the false negatives were cleared out.

We've stopped trying to delete things that don't exist, which was causing most of the false negatives

Upon checking to see if an ID exists in a scope, we now check to see what the ID's type is, so that we can compare it to things. This was causing the rest of the false negatives.

Semantic Analysis has been added back to lexAndParse/1, along with a message to acknowledge that semantic analysis has completed before presenting the symbol table (which was also added).

Changed the failure state of the semantic analysis to immediately stop upon encountering an error. Asking it to keep going resulted in an extremely long attempt to correct the issue that led it to stop, which likely would have eventually resulted in it switching from one scope to another or something like that, but regardless, it took an inordinate amount of time.